requirements.txt
Flask
tensorflow
werkzeug
numpy
pillow